+++ changelog.md
@@ -3,14 +3,16 @@
 ## v0.17.5
 
 - **New:**
-  = Add [%escape(string)](docs/functions.md#string)  escape all characters in string if escaping is allowed and if those characters need escaping
+  - Add [%escape(string)](docs/functions.md#string)  escape all characters in string if escaping is allowed and if those characters need escaping
   - Add [%unsxape(string)](docs/functions.md#string) strip escape characters from string if escaping is allowed and those characters are enabled
+  - Add $selectedword, $selword, $selectedurl, $selurl, $selectedline, $selline, $selected, $character, $copied for direct access in expressions
 - **Fixed:**
   - Input parser:
     - Correctly escape function arguments when quoting arguments with respect to scripting quote settings
     - Use custom escape character when splitting function arguments with quotes
   - Fix dice when using fudge dice
   - Fix parsing of sub context menus
+  - Fix $selectedword, $selword, $selectedurl, $selurl, $selectedline, $selline, when used from command line to use mouse position to correctly get the selected word,line, or url
 - **Changed:**
   - Error messages for commands and functions will add name to better understand where the error is